refactor: move all referencies to database into one folder in infra layer
