+++ src/infra/database/index.ts
@@ -1,2 +1,3 @@
 export * from './contracts'
-export * from './adapters'
+export * from './models'
+export * from './postgres'
