+++ src/infra/database/models/item/ItemTable.ts
