+++ src/infra/database/models/item/types/index.ts
