+++ src/infra/database/models/item/types/itemTable.ts
