+++ src/infra/database/postgres/index.ts
@@ -1,0 +1,2 @@
+export * from './adapters'
+export * from './repositories'
