+++ src/infra/database/postgres/repositories/item/ItemRepositoryPostgres.ts
@@ -2,8 +2,7 @@ import { Item } from '@/domain/entities'
 import { ItemRepository } from '@/domain/repositories'
 import { ItemNotFoundError } from '@/domain/errors'
 
-import { ItemTable } from '@/infra/models'
-import { DatabaseConnection } from '@/infra/database'
+import { DatabaseConnection, ItemTable } from '@/infra/database'
 
 export class ItemRepositoryPostgres implements ItemRepository {
   private readonly databaseConnection: DatabaseConnection
