--- src/infra/repositories/index.ts
@@ -1,1 +1,0 @@
-export * from './postgres'
